+++ app/src/main/java/com/dabenxiang/mimi/model/api/ApiService.kt
@@ -204,10 +204,10 @@ interface ApiService {
     ): Response<ApiBasePagingItem<List<MemberClubItem>>>
 
     @POST("/v1/Members/Club/{clubId}/Follow")
-    suspend fun followClub(@Path("clubId") clubId: Int): Response<Void>
+    suspend fun followClub(@Path("clubId") clubId: Long): Response<Void>
 
     @DELETE("/v1/Members/Club/{clubId}/Follow")
-    suspend fun cancelFollowClub(@Path("clubId") clubId: Int): Response<Void>
+    suspend fun cancelFollowClub(@Path("clubId") clubId: Long): Response<Void>
 
     /**********************************************************
      *
